+++ src/main/java/com/batsworks/matrix/factory/Operation.java
@@ -3,7 +3,7 @@ package com.batsworks.matrix.factory;
 import com.batsworks.matrix.enums.OperationEnum;
 
 public interface Operation {
-    int[][] execute(int[][] matA, int[][] matB, int cols, int rows);
+    double[][] execute(double[][] matA, double[][] matB, int cols, int rows);
 
     boolean use(OperationEnum operationEnum);
 }
